requirements.txt
arrow==1.2.3
asgiref==3.6.0
certifi==2023.5.7
click==8.1.3
colorama==0.4.6
Django==4.2.1
djangorestframework==3.14.0
geographiclib==2.0
pip==21.1.1
proj==0.2.0
psycopg2==2.9.6
pyproj==3.5.0
python-dateutil==2.8.2
pytz==2023.3
setuptools==56.0.0
six==1.16.0
sqlparse==0.4.4
tzdata==2023.3
utm==0.7.0